+++ src/main/java/plugily/projects/murdermystery/commands/arguments/ArgumentsRegistry.java
@@ -32,6 +32,7 @@ import pl.plajerlair.commonsbox.string.StringMatcher;
 import plugily.projects.murdermystery.Main;
 import plugily.projects.murdermystery.arena.ArenaRegistry;
 import plugily.projects.murdermystery.commands.arguments.admin.ListArenasArgument;
+import plugily.projects.murdermystery.commands.arguments.admin.arena.SpecialBlockRemoverArgument;
 import plugily.projects.murdermystery.commands.arguments.admin.arena.DeleteArgument;
 import plugily.projects.murdermystery.commands.arguments.admin.arena.ForceStartArgument;
 import plugily.projects.murdermystery.commands.arguments.admin.arena.ReloadArgument;
@@ -90,6 +91,7 @@ public class ArgumentsRegistry implements CommandExecutor {
     new DeleteArgument(this, chatManager);
     new ForceStartArgument(this, chatManager);
     new ListArenasArgument(this, chatManager);
+    new SpecialBlockRemoverArgument(this);
     new ReloadArgument(this, chatManager);
     new StopArgument(this);
   }