+++ src/main/java/plugily/projects/murdermystery/utils/Utils.java
@@ -120,7 +120,7 @@ public class Utils {
   }
 
   public static Location getBlockCenter(Location location) {
-    return location.add(0.5, 0, 0.5);
+    return location.clone().add(0.5, 0, 0.5);
   }
 
   public static boolean checkIsInGameInstance(Player player) {